+++ lib/new_relic/agent/instrumentation/active_record_5.rb
@@ -1,64 +1,9 @@
 # encoding: utf-8
 # This file is distributed under New Relic's license terms.
 # See https://github.com/newrelic/rpm/blob/master/LICENSE for complete details.
+
 require 'new_relic/agent/instrumentation/active_record_subscriber'
-require 'new_relic/agent/prepend_supportability'
-
-module NewRelic
-  module Agent
-    module Instrumentation
-      module ActiveRecord
-        ACTIVE_RECORD = "ActiveRecord".freeze
-
-        module BaseExtensions
-          def save(*args, &blk)
-            ::NewRelic::Agent.with_database_metric_name(self.class.name, nil, ACTIVE_RECORD) do
-              super
-            end
-          end
-
-          def save!(*args, &blk)
-            ::NewRelic::Agent.with_database_metric_name(self.class.name, nil, ACTIVE_RECORD) do
-              super
-            end
-          end
-        end
-
-        module RelationExtensions
-          def update_all(*args, &blk)
-            ::NewRelic::Agent.with_database_metric_name(self.name, nil, ACTIVE_RECORD) do
-              super
-            end
-          end
-
-          def delete_all(*args, &blk)
-            ::NewRelic::Agent.with_database_metric_name(self.name, nil, ACTIVE_RECORD) do
-              super
-            end
-          end
-
-          def destroy_all(*args, &blk)
-            ::NewRelic::Agent.with_database_metric_name(self.name, nil, ACTIVE_RECORD) do
-              super
-            end
-          end
-
-          def calculate(*args, &blk)
-            ::NewRelic::Agent.with_database_metric_name(self.name, nil, ACTIVE_RECORD) do
-              super
-            end
-          end
-
-          def pluck(*args, &blk)
-            ::NewRelic::Agent.with_database_metric_name(self.name, nil, ACTIVE_RECORD) do
-              super
-            end
-          end
-        end
-      end
-    end
-  end
-end
+require 'new_relic/agent/instrumentation/active_record_prepend'
 
 DependencyDetection.defer do
   named :active_record_5
@@ -85,8 +30,8 @@ DependencyDetection.defer do
       NewRelic::Agent::Instrumentation::ActiveRecordSubscriber.new)
 
     ActiveSupport.on_load(:active_record) do
-      ::ActiveRecord::Base.prepend ::NewRelic::Agent::Instrumentation::ActiveRecord::BaseExtensions
-      ::ActiveRecord::Relation.prepend ::NewRelic::Agent::Instrumentation::ActiveRecord::RelationExtensions
+      ::ActiveRecord::Base.prepend ::NewRelic::Agent::Instrumentation::ActiveRecordPrepend::BaseExtensions
+      ::ActiveRecord::Relation.prepend ::NewRelic::Agent::Instrumentation::ActiveRecordPrepend::RelationExtensions
     end
   end
 end